requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
arrow==1.2.2
bbcode==1.1.0
deep-translator==1.8.3
emoji==1.7.0
feedparser==6.0.10
fuzzyfinder==2.1.0
httpx==0.23.0
imagehash==4.2.1
imageio==2.19.3
jieba==0.42.1
magneturi==1.3.0
nonebot-adapter-onebot==2.1.1
nonebot-plugin-apscheduler==0.1.3
nonebot-plugin-docs==2.0.0b4
nonebot-plugin-gocqhttp==0.5.5
nonebot-plugin-guild-patch==0.2.0
nonebot-plugin-htmlrender==0.0.5
nonebot-plugin-imageutils==0.1.6.3
nonebot-plugin-status==0.4.0
nonebot2==2.0.0b4
pyppeteer==1.0.2
pyquery==1.4.3
qbittorrent==0.1.6
qrcode==7.3.1
tenacity=8.0.1
tinydb==4.7.0
ujson==5.3.0
wordcloud==1.8.1